test
Using various textures to build up the final render in a shader.

There's issues with the output of the shader being completely black, even if add() is being used. Need to investigate that.

Using various textures to build up the final render in a shader.

There's issues with the output of the shader being completely black, even if add() is being used. Need to investigate that.

Using various textures to build up the final render in a shader.

There's issues with the output of the shader being completely black, even if add() is being used. Need to investigate that.

test

Using various textures to build up the final render in a shader.

There's issues with the output of the shader being completely black, even if add() is being used. Need to investigate that.